requirements.txt
# Core dependencies
pygame-ce>=2.5.0
numpy>=1.24.0

# Data handling
toml>=0.10.2

# Performance monitoring (for debug overlay)
psutil>=5.9.0

# Optional dependencies for development
# pytest>=7.4.0
# pytest-cov>=4.1.0
# black>=23.0.0
# pylint>=2.17.0
# mypy>=1.5.0

# Build tools (optional)
# pyinstaller>=5.13.0  # For creating executables
# cx_Freeze>=6.15.0   # Alternative for creating executables